Fix compilation errors on Alpine Linux
mxs_strerror macro also removes the need to always define a local error
buffer by using the thread-safe function strerror_l.

The thread_self macro should be used instead of calling pthread_self
directly. Since it's never used to access the thread object, the macro
can expand to a size_t instead of a THREAD.

Add some missing definitions which aren't defined on other platforms.

Rename maxscale/poll.h header guards as they conflicted with the ones in
sys/poll.h.

+++ server/include/thread.h
@@ -26,8 +26,16 @@
  * Thread type and thread identifier function macros
  */
 #include <pthread.h>
+
+/** Thread type */
 #define THREAD         pthread_t
-#define thread_self()  pthread_self()
+
+/** Thread identity
+ *
+ * This is never used to access the thread object but only to identify which
+ * thread is active. This information is almost always logged so returning
+ * an integer makes more sense. */
+#define thread_self()  (size_t)pthread_self()
 
 extern THREAD *thread_start(THREAD *thd, void (*entry)(void *), void *arg);
 extern void thread_wait(THREAD thd);
